src/whisper-to-text/requirements.txt
pyaudio
pyperclip
openai-whisper
pywhispercpp
